refactor(ui): split LayoutStoreModal.tsx into 5 focused modules (795→42 lines)
Extract types, Hub actions, entry card, and content from LayoutStoreModal.

Modules extracted:
- layout-store-types.ts: FileStatus, HubEntryResult, LayoutStoreContentProps, button constants
- LayoutStoreHubActions.tsx: HubOrphanButtons, ShareLink, LayoutStoreHubRow
- LayoutStoreEntry.tsx: FormatButtons, per-entry card component
- LayoutStoreContent.tsx: save form, history list, import section
- LayoutStoreModal.tsx: thin modal shell with backward-compat re-exports
